+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - ... <!-- Add new lines here. -->
 - feat: Add component `Radio`
+- feat: Add `--nolabel` variant for `Checkbox`
 
 ## 0.8.36