+++ modules/hanna-react/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - ... <!-- Add new lines here. -->
 - feat: Add component `Radio` as a standalone atomic widget
+- feat: Add prop `hideLabel` to `Checkbox` for minimal/atomic use
 - fix: Regression in `Checkbox`- and `Radio`-related option typing
 
 ## 0.10.113